move enterAuthFlow logic to OpenAuthoring
