+++ open-authoring/open-authoring/OpenAuthoringProvider.tsx
@@ -1,9 +1,8 @@
 import React, { useState, useEffect } from 'react'
-import { getForkName, getHeadBranch } from './repository'
+import { getForkName, getHeadBranch, setForkName } from './repository'
 import { useCMS } from 'tinacms'
 import { SourceProviderManager } from '../source-provider-managers/SourceProviderManager'
-import { enterEditMode } from '../auth/authFlow'
-
+import popupWindow from '../../utils/popupWindow'
 export interface OpenAuthoringContext {
   forkValid: boolean
   githubAuthenticated: boolean
@@ -48,7 +47,31 @@ export const OpenAuthoringProvider = ({
   }, [])
 
   const edit = () => {
-    enterEditMode(githubAuthenticated, forkValid)
+    const fork = getForkName()
+
+    if (githubAuthenticated) {
+      if (fork && forkValid) {
+        setForkName(fork)
+        fetch(`/api/preview`).then(() => {
+          window.location.href = window.location.pathname
+        })
+        return
+      } else {
+        popupWindow(`/github/fork`, '_blank', window, 1000, 700)
+      }
+    } else {
+      const authState = Math.random()
+        .toString(36)
+        .substring(7)
+
+      popupWindow(
+        `/github/start-auth?state=${authState}`,
+        '_blank',
+        window,
+        1000,
+        700
+      )
+    }
   }
 
   return (
