+++ utils/popupWindow.ts
@@ -1,0 +1,18 @@
+function popupWindow(url, title, window, w, h) {
+  const y = window.top.outerHeight / 2 + window.top.screenY - h / 2
+  const x = window.top.outerWidth / 2 + window.top.screenX - w / 2
+  return window.open(
+    url,
+    title,
+    'toolbar=no, location=no, directories=no, status=no, menubar=no, scrollbars=yes, resizable=yes, copyhistory=no, width=' +
+      w +
+      ', height=' +
+      h +
+      ', top=' +
+      y +
+      ', left=' +
+      x
+  )
+}
+
+export default popupWindow
